Bug 1165006 - Remove Profile argument from NSUserDefaultsProfilePrefs and rename to NSUserDefaultsPrefs
